requirements.txt
fastapi
uvicorn
firebase-admin
requests
python-dotenv
